+++ src/main/java/ro/ebob/oe/repo/OrderRepository.java
@@ -5,5 +5,5 @@ import ro.ebob.oe.repo.entity.Order;
 
 import java.math.BigDecimal;
 
-public interface OrderRepository extends JpaRepository<Order, BigDecimal> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
